add flask; add catboost with annoy

+++ src/test_utils.py
@@ -14,7 +14,8 @@ from private.settings import PRIVATE_DIR
 from src.vectors import DoctorVector
 from src.db_func import (
     get_users, get_last_doctor_with_all_in_town, get_all_appointment_users,
-    get_town_doctor_list, get_all_appointments, get_two_last_appointment,
+    get_town_doctor_list, get_all_appointments, get_two_last_appointment, get_all_users, get_all_last_appointment,
+    get_last_doctor_with_all_in_town_with_all_appts,
 )
 
 
@@ -111,6 +112,48 @@ def create_train_data(min_appts: int) -> Tuple[pd.DataFrame, pd.DataFrame]:
     return test_df, train_df
 
 
+def create_top_one_train_data(min_appts: int) -> Tuple[pd.DataFrame, pd.DataFrame]:
+    session = Session()
+    all_users = get_all_appointment_users(session, min_appts)
+
+    doctor_vectors = DoctorVector()
+    doctor_vectors.load()
+
+    test = []
+    train = []
+
+    for user in progressbar(all_users):
+        # add test data
+        last_appt, test_appt = get_two_last_appointment(session, user)
+
+        test_doctor_model = session.query(Doctor).filter(Doctor.id == test_appt.doctor_id).first()
+        test_doctors = get_town_doctor_list(session, test_doctor_model.town_id, test_appt.spec_id)
+        test_appt_id = test_appt.id
+        for doctor in test_doctors[:100]:
+            if doctor == test_doctor_model.id:
+                continue
+            test.append([0, test_appt_id, *doctor_vectors[doctor]])
+        test.append([1, test_appt_id, *doctor_vectors[test_doctor_model.id]])
+
+        # add train data
+        all_user_appts = get_all_appointments(session, user, [last_appt.doctor_id, test_appt.doctor_id])
+        for appt in all_user_appts:
+            appt_id = appt.id
+            train.append([1, appt_id, *doctor_vectors[appt.doctor_id]])
+
+            train_doctor_model = session.query(Doctor).filter(Doctor.id == appt.doctor_id).first()
+            train_town_doctors = get_town_doctor_list(session, train_doctor_model.town_id, appt.spec_id)
+            for doctor in train_town_doctors[:100]:
+                if doctor == train_doctor_model.id:
+                    continue
+                train.append([0, appt_id, *doctor_vectors[doctor]])
+
+    test_df = pd.DataFrame(test)
+    train_df = pd.DataFrame(train)
+
+    return test_df, train_df
+
+
 def get_train_data(min_appts=0, flush=False):
     train_path = os.path.join(PRIVATE_DIR, 'ranking', f'train_data_{min_appts}.csv')
     test_path = os.path.join(PRIVATE_DIR, 'ranking', f'test_data_{min_appts}.csv')
@@ -120,6 +163,21 @@ def get_train_data(min_appts=0, flush=False):
         test_df = pd.read_csv(test_path, header=None)
     else:
         test_df, train_df = create_train_data(min_appts)
+        test_df.to_csv(test_path, index=False, header=False)
+        train_df.to_csv(train_path, index=False, header=False)
+
+    return test_df, train_df
+
+
+def get_top_one_train_data(min_appts=0, flush=False):
+    train_path = os.path.join(PRIVATE_DIR, 'ranking', f'top_one_train_data_{min_appts}.csv')
+    test_path = os.path.join(PRIVATE_DIR, 'ranking', f'top_one_test_data_{min_appts}.csv')
+
+    if os.path.exists(train_path) and os.path.exists(test_path) and not flush:
+        train_df = pd.read_csv(train_path, header=None)
+        test_df = pd.read_csv(test_path, header=None)
+    else:
+        test_df, train_df = create_top_one_train_data(min_appts)
         test_df.to_csv(test_path, index=False, header=False)
         train_df.to_csv(train_path, index=False, header=False)
 
@@ -138,13 +196,42 @@ def get_test_data(min_appt=5):
 
     user_data = {}
     for user in progressbar(users):
-        last_doctor, town_doctors, all_appts = get_last_doctor_with_all_in_town(session, user)
+        last_doctor, town_doctors, all_appts = get_last_doctor_with_all_in_town_with_all_appts(session, user)
         if last_doctor.id not in town_doctors:
             continue
         user_data[user] = {
             'selected_doctor': last_doctor.id,
             'suggested_doctors': town_doctors,
             'all_appts': [appt.doctor_id for appt in all_appts],
+        }
+
+    with open(file_path, 'w') as f:
+        json.dump(user_data, f)
+    print('len ->', len(user_data))
+    return user_data
+
+
+def get_biggest_data():
+    file_path = os.path.join(PRIVATE_DIR, 'vectors', f'test_full_data.json')
+
+    if os.path.exists(file_path):
+        with open(file_path, 'r') as f:
+            return json.load(f)
+
+    session = Session()
+    appts = get_all_last_appointment(session)
+
+    user_data = {}
+    for appt in progressbar(appts, max_value=appts.count()):
+        user_id, appt_id, spec_id, town_id, last_doctor_id = appt
+        town_doctors = get_town_doctor_list(session, town_id, spec_id)
+
+        if last_doctor_id not in town_doctors:
+            continue
+        user_data[user_id] = {
+            'selected_doctor': last_doctor_id,
+            'suggested_doctors': town_doctors,
+            'all_appts': [],
         }
 
     with open(file_path, 'w') as f:
